changed structure and added tests
